refactor: split metre into mod, prediction, classical placeholder
